# src/lib/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Horizon IT Management API",
    "version": "1.0.0",
    "description": "API for managing IT tickets, assets, and users."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Internal API"
    }
  ],
  "paths": {
    "/tickets": {
      "get": {
        "summary": "List tickets",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": { "type": "integer", "default": 20 }
          },
          {
            "name": "all",
            "in": "query",
            "schema": { "type": "boolean", "default": false }
          }
        ],
        "responses": {
          "200": {
            "description": "Output is a paginated list of tickets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tickets": { "type": "array", "items": { "$ref": "#/components/schemas/Ticket" } },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": { "type": "integer" },
                        "totalPages": { "type": "integer" },
                        "totalCount": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create ticket",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TicketCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "Ticket created" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Ticket": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string" },
          "priority": { "type": "string" },
          "requesterName": { "type": "string", "nullable": true }
        }
      },
      "TicketCreate": {
        "type": "object",
        "required": ["title", "description", "status", "priority"],
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string" },
          "priority": { "type": "string" },
          "assignedToId": { "type": "integer", "nullable": true },
          "assetId": { "type": "integer", "nullable": true }
        }
      }
    }
  }
}